# APIServer/Areas/Content/Models/Album.cs
using APIServer.Utilities;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace APIServer.Areas.Content.Models
{
    public class Album : BaseEntity
    {
        public int Id { get; set; }
        public string Title { get; set; }
        public string Genre { get; set; }
        public ICollection<Song> Songs { get; set; }
        public string? CoverImage { get; set; }
    }
}