# APIServer/Areas/Content/Models/Post.cs
using APIServer.Utilities;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace APIServer.Areas.Content.Models
{
    public class Post : BaseEntity
    {
        public int Id { get; set; }
        public string Title { get; set; }
        public string? Content { get; set; }
        public string? Link { get; set; }
        public ICollection<Media>? Gallery { get; set; }
    }
}