# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "APIServer",
    "version": "v1"
  },
  "paths": {
...
    "/Articles": {
      "get": {
        "tags": [
          "Articles"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Articles"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          }
        }
      }
    },
    "/Articles/{id}": {
      "get": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/Media": {
      "get": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Media"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            }
          }
        }
      }
    },
    "/Media/{id}": {
      "get": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Song": {
        "required": [
          "album",
          "title"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "album": {
            "$ref": "#/components/schemas/Album"
          },
          "title": {
            "type": "string"
          },
          "genre": {
            "type": "string",
            "nullable": true
          },
          "soundCloud": {
            "type": "string",
            "nullable": true
          },
          "audio": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Album": {
        "required": [
          "genre",
          "title"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Song"
            },
            "nullable": true
          },
          "coverImage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Post": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "link": {
            "type": "string",
            "nullable": true
          },
          "gallery": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Media"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MediaType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "Media": {
        "required": [
          "type",
          "url"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "post": {
            "$ref": "#/components/schemas/Post"
          },
          "article": {
            "$ref": "#/components/schemas/Article"
          },
          "url": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/MediaType"
          }
        },
        "additionalProperties": false
      },
      "Article": {
        "required": [
          "content",
          "title"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "song": {
            "$ref": "#/components/schemas/Song"
          },
          "gallery": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Media"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}